requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-aiplatform==1.70.0
python-dotenv==1.0.1
python-multipart==0.0.12
